refactor: move linkage to common (#27)
