+++ example/src/config/index.js
@@ -9,6 +9,7 @@ export const SCOPE_MAP = {
   // '@tuya/tuya-panel-health-sdk': 'Health',
   '@tuya/tuya-panel-ipc-sdk': 'Ipc',
   '@tuya/tuya-panel-lamp-sdk': 'Lamp',
+  '@tuya/tuya-panel-api': 'ApiScene',
   // '@tuya/tuya-panel-remote-sdk': 'Remote',
   // '@tuya/tuya-panel-robot-sdk': 'SweepRobot',
   // '@tuya/tuya-panel-sensor-sdk': 'Sensor',
