+++ packages/tuya-panel-api/src/common/linkage.ts
