--- packages/tuya-panel-api/src/electrician/index.ts
@@ -1,5 +1,0 @@
-import linkageApi from './linkage';
-
-export default {
-  linkageApi,
-};
