--- packages/tuya-panel-api/src/electrician/interface.ts
@@ -1,25 +1,0 @@
-export interface IGetLinkageDeviceList {
-  gid: number;
-  sourceType: string;
-}
-
-export interface IBindRule {
-  associativeEntityId: string;
-  ruleId: string;
-  entitySubIds: string;
-  expr: [string[]];
-  bizDomain: string;
-}
-
-export interface IGetBindRuleList {
-  bizDomain: string;
-  sourceEntityId: string;
-  entityType: number;
-}
-
-export interface IRemoveRule {
-  bizDomain: string;
-  sourceEntityId: string;
-  associativeEntityId: string;
-  associativeEntityValue: string;
-}
