+++ packages/tuya-panel-api/src/index.ts
@@ -1,7 +1,5 @@
 import common from './common';
 import robot from './robot';
-import electrician from './electrician';
 
 export const commonApi = common;
 export const robotApi = robot;
-export const electricianApi = electrician;
